+++ src/org/schema/constants/traits/FlightReservation.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The FlightReservation properties enumeration.
+ */
+trait FlightReservation
+{
+    const string BOARDING_GROUP             = 'boardingGroup'            ;
+    const string PASSENGER_PRIORITY_STATUS  = 'passengerPriorityStatus'  ;
+    const string PASSENGER_SEQUENCE_NUMBER  = 'passengerSequenceNumber'  ;
+    const string SECURITY_SCREENING         = 'securityScreening'        ;
+}