+++ src/org/schema/constants/traits/FoodEstablishmentReservation.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The FoodEstablishmentReservation properties enumeration.
+ */
+trait FoodEstablishmentReservation
+{
+    const string END_TIME   = 'endTime'   ;
+    const string PARTY_SIZE = 'partySize' ;
+    const string START_TIME = 'startTime' ;
+}