+++ src/org/schema/constants/traits/LodgingReservation.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The LodgingReservation properties enumeration.
+ */
+trait LodgingReservation
+{
+    const string CHECKIN_TIME             = 'checkinTime'            ;
+    const string CHECKOUT_TIME            = 'checkoutTime'           ;
+    const string LODGING_UNIT_DESCRIPTION = 'lodgingUnitDescription' ;
+    const string LODGING_UNIT_TYPE        = 'lodgingUnitType'        ;
+    const string NUM_ADULTS               = 'numAdults'              ;
+    const string NUM_CHILDREN             = 'numChildren'            ;
+}