+++ src/org/schema/constants/traits/RentalCarReservation.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The RentalCarReservation properties enumeration.
+ */
+trait RentalCarReservation
+{
+    const string DROPOFF_LOCATION = 'dropoffLocation' ;
+    const string DROPOFF_TIME     = 'dropoffTime'     ;
+    const string PICKUP_LOCATION  = 'pickupLocation'  ;
+    const string PICKUP_TIME      = 'pickupTime'      ;
+}