+++ src/org/schema/constants/traits/Reservation.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The Reservation properties enumeration.
+ */
+trait Reservation
+{
+    const string BOOKING_TIME            = 'bookingTime'           ;
+    const string BROKER                  = 'broker'                ;
+    const string MODIFIED_TIME           = 'modifiedTime'          ;
+    const string PRICE_CURRENCY          = 'priceCurrency'         ;
+    const string PROGRAM_MEMBERSHIP_USED = 'programMembershipUsed' ;
+    const string PROVIDER                = 'provider'              ;
+    const string RESERVATION_FOR         = 'reservationFor'        ;
+    const string RESERVATION_ID          = 'reservationId'         ;
+    const string RESERVATION_STATUS      = 'reservationStatus'     ;
+    const string RESERVED_TICKET         = 'reservedTicket'        ;
+    const string TOTAL_PRICE             = 'totalPrice'            ;
+    const string UNDER_NAME              = 'underName'             ;
+}