+++ src/org/schema/constants/traits/ReservationPackage.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The ReservationPackage properties enumeration.
+ */
+trait ReservationPackage
+{
+    const string SUB_RESERVATION = 'subReservation' ;
+}