+++ src/org/schema/constants/traits/TaxiReservation.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The TaxiReservation properties enumeration.
+ */
+trait TaxiReservation
+{
+    const string PARTY_SIZE      = 'partySize'      ;
+    const string PICKUP_LOCATION = 'pickupLocation' ;
+    const string PICKUP_TIME     = 'pickupTime'     ;
+}