+++ src/org/schema/constants/traits/Ticket.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The Ticket properties enumeration.
+ */
+trait Ticket
+{
+    const string DATE_ISSUED    = 'dateIssued'    ;
+    const string ISSUED_BY      = 'issuedBy'      ;
+    const string PRICE_CURRENCY = 'priceCurrency' ;
+    const string TICKET_NUMBER  = 'ticketNumber'  ;
+    const string TICKET_TOKEN   = 'ticketToken'   ;
+    const string TICKETED_SEAT  = 'ticketedSeat'  ;
+    const string TOTAL_PRICE    = 'totalPrice'    ;
+    const string UNDER_NAME     = 'underName'     ;
+}